+++ src/main.cpp
@@ -7,6 +7,7 @@ $execute
 	new geode::EventListener(+[](geode::GameEvent*) {
 		ng::globals::g_isEditorIDAPILoaded = geode::Loader::get()->isModLoaded("cvolton.level-id-api");
 		ng::globals::g_isBetterEditLoaded = geode::Loader::get()->isModLoaded("hjfod.betteredit");
+		ng::globals::g_isBetterColorPickerLoaded = geode::Loader::get()->isModLoaded("anatom3000.bettercolorpicker");
 
 
 		ng::globals::g_buildHelperRawNameFormat = geode::Mod::get()->getSettingValue<std::string>("auto-name-format");